refactor(frontend): centralize cache clearing utility
