+++ frontend/src/components/pwa/CacheManager.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { Button } from "../ui/button";
 import { CACHE_VERSION } from "@/config/pwa";
-import { clearCache } from "@/services/admin/cacheService";
+import { clearCache } from "@/utils/cache";
 
 // List of URLs to warm up in cache. Replace with actual routes as needed.
 const pwaWarmList: string[] = [];
