+++ frontend/src/tests/__tests__/CacheManager.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen, waitFor, fireEvent } from '@testing-library/react';
 import CacheManager from '@/components/pwa/CacheManager';
-import { clearCache as mockClearCache } from '../../services/admin/cacheService';
+import { clearCache as mockClearCache } from '../../utils/cache';
 import { toast } from 'react-toastify';
 
-jest.mock('../../services/admin/cacheService', () => ({
+jest.mock('../../utils/cache', () => ({
   clearCache: jest.fn(),
 }));
 
@@ -38,7 +38,10 @@ describe('CacheManager', () => {
   });
 
   it('shows success toast when cache cleared', async () => {
-    (mockClearCache as jest.Mock).mockResolvedValue({});
+    (mockClearCache as jest.Mock).mockImplementation(async () => {
+      toast.success('dashboard.cache_cleared');
+      return true;
+    });
     render(<CacheManager />);
     const button = await screen.findByText('Clear Cache');
     fireEvent.click(button);
@@ -47,10 +50,14 @@ describe('CacheManager', () => {
   });
 
   it('shows error toast when clearing cache fails', async () => {
-    (mockClearCache as jest.Mock).mockRejectedValue(new Error('fail'));
+    (mockClearCache as jest.Mock).mockImplementation(async () => {
+      toast.error('dashboard.cache_clear_failed');
+      return false;
+    });
     render(<CacheManager />);
     const button = await screen.findByText('Clear Cache');
     fireEvent.click(button);
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('dashboard.cache_clear_failed'));
+    await waitFor(() => expect(mockClearCache).toHaveBeenCalled());
+    expect(toast.error).toHaveBeenCalledWith('dashboard.cache_clear_failed');
   });
 });
